+++ test/abstract/2.sym-cast-f.cpp
@@ -15,7 +15,7 @@ int main() {
         short y;
     };
     x = nondet();
-    if ( x < 0 )
+    if ( short( x ) < 0 )
         return 0;
     ++y;
     assert( y == -32768 || y >= 1 );
